(#20) Split keypress.c
