+++ src/macos/keypress.c
@@ -1,34 +1,13 @@
-#include "keypress.h"
-#include "deadbeef_rand.h"
-#include "microsleep.h"
+#include "../keypress.h"
+#include "../deadbeef_rand.h"
+#include "../microsleep.h"
 
 #include <ctype.h> /* For isupper() */
 
-#if defined(IS_MACOSX)
-	#include <ApplicationServices/ApplicationServices.h>
-	#import <IOKit/hidsystem/IOHIDLib.h>
-	#import <IOKit/hidsystem/ev_keymap.h>
-#elif defined(USE_X11)
-	#include <X11/extensions/XTest.h>
-	#include "xdisplay.h"
-#endif
+#include <ApplicationServices/ApplicationServices.h>
+#import <IOKit/hidsystem/IOHIDLib.h>
+#import <IOKit/hidsystem/ev_keymap.h>
 
-/* Convenience wrappers around ugly APIs. */
-#if defined(IS_WINDOWS)
-	#define WIN32_KEY_EVENT_WAIT(key, flags) \
-		(win32KeyEvent(key, flags), Sleep(DEADBEEF_RANDRANGE(0, 63)))
-#elif defined(USE_X11)
-	#define X_KEY_EVENT(display, key, is_press) \
-		(XTestFakeKeyEvent(display, \
-		                   XKeysymToKeycode(display, key), \
-		                   is_press, CurrentTime), \
-		 XSync(display, false))
-	#define X_KEY_EVENT_WAIT(display, key, is_press) \
-		(X_KEY_EVENT(display, key, is_press), \
-		 microsleep(DEADBEEF_UNIFORM(0.0, 62.5)))
-#endif
-
-#if defined(IS_MACOSX)
 static io_connect_t _getAuxiliaryKeyDriver(void)
 {
 	static mach_port_t sEventDrvrRef = 0;
@@ -49,71 +28,9 @@ static io_connect_t _getAuxiliaryKeyDriver(void)
 	}
 	return sEventDrvrRef;
 }
-#endif
-
-#if defined(IS_WINDOWS)
-void win32KeyEvent(int key, MMKeyFlags flags)
-{
-	int scan = MapVirtualKey(key & 0xff, MAPVK_VK_TO_VSC);
-
-	/* Set the scan code for extended keys */
-	switch (key)
-	{
-		case VK_RCONTROL:
-		case VK_SNAPSHOT: /* Print Screen */
-		case VK_RMENU: /* Right Alt / Alt Gr */
-		case VK_PAUSE: /* Pause / Break */
-		case VK_HOME:
-		case VK_UP:
-		case VK_PRIOR: /* Page up */
-		case VK_LEFT:
-		case VK_RIGHT:
-		case VK_END:
-		case VK_DOWN:
-		case VK_NEXT: /* 'Page Down' */
-		case VK_INSERT:
-		case VK_DELETE:
-		case VK_LWIN:
-		case VK_RWIN:
-		case VK_APPS: /* Application */
-		case VK_VOLUME_MUTE:
-		case VK_VOLUME_DOWN:
-		case VK_VOLUME_UP:
-		case VK_MEDIA_NEXT_TRACK:
-		case VK_MEDIA_PREV_TRACK:
-		case VK_MEDIA_STOP:
-		case VK_MEDIA_PLAY_PAUSE:
-		case VK_BROWSER_BACK:
-		case VK_BROWSER_FORWARD:
-		case VK_BROWSER_REFRESH:
-		case VK_BROWSER_STOP:
-		case VK_BROWSER_SEARCH:
-		case VK_BROWSER_FAVORITES:
-		case VK_BROWSER_HOME:
-		case VK_LAUNCH_MAIL:
-		{
-			flags |= KEYEVENTF_EXTENDEDKEY;
-			break;
-		}
-	}
-
-	/* Set the scan code for keyup */
-	if ( flags & KEYEVENTF_KEYUP ) {
-		scan |= 0x80;
-	}
-
-	INPUT keyboardInput;
-	keyboardInput.type = INPUT_KEYBOARD;
-	keyboardInput.ki.wScan = scan;
-	keyboardInput.ki.dwFlags = KEYEVENTF_SCANCODE | flags;
-	keyboardInput.ki.time = 0;
-	SendInput(1, &keyboardInput, sizeof(keyboardInput));
-}
-#endif
 
 void toggleKeyCode(MMKeyCode code, const bool down, MMKeyFlags flags)
 {
-#if defined(IS_MACOSX)
 	/* The media keys all have 1000 added to them to help us detect them. */
 	if (code >= 1000) {
 		code = code - 1000; /* Get the real keycode. */
@@ -127,8 +44,7 @@ void toggleKeyCode(MMKeyCode code, const bool down, MMKeyFlags flags)
 		kr = IOHIDPostEvent( _getAuxiliaryKeyDriver(), NX_SYSDEFINED, loc, &event, kNXEventDataVersion, 0, FALSE );
 		assert( KERN_SUCCESS == kr );
 	} else {
-		CGEventRef keyEvent = CGEventCreateKeyboardEvent(NULL,
-		                                                 (CGKeyCode)code, down);
+		CGEventRef keyEvent = CGEventCreateKeyboardEvent(NULL, (CGKeyCode)code, down);
 		assert(keyEvent != NULL);
 
 		CGEventSetType(keyEvent, down ? kCGEventKeyDown : kCGEventKeyUp);
@@ -136,28 +52,6 @@ void toggleKeyCode(MMKeyCode code, const bool down, MMKeyFlags flags)
 		CGEventPost(kCGSessionEventTap, keyEvent);
 		CFRelease(keyEvent);
 	}
-#elif defined(IS_WINDOWS)
-	const DWORD dwFlags = down ? 0 : KEYEVENTF_KEYUP;
-
-	/* Parse modifier keys. */
-	if (flags & MOD_META) WIN32_KEY_EVENT_WAIT(K_META, dwFlags);
-	if (flags & MOD_ALT) WIN32_KEY_EVENT_WAIT(K_ALT, dwFlags);
-	if (flags & MOD_CONTROL) WIN32_KEY_EVENT_WAIT(K_CONTROL, dwFlags);
-	if (flags & MOD_SHIFT) WIN32_KEY_EVENT_WAIT(K_SHIFT, dwFlags);
-
-	win32KeyEvent(code, dwFlags);
-#elif defined(USE_X11)
-	Display *display = XGetMainDisplay();
-	const Bool is_press = down ? True : False; /* Just to be safe. */
-
-	/* Parse modifier keys. */
-	if (flags & MOD_META) X_KEY_EVENT_WAIT(display, K_META, is_press);
-	if (flags & MOD_ALT) X_KEY_EVENT_WAIT(display, K_ALT, is_press);
-	if (flags & MOD_CONTROL) X_KEY_EVENT_WAIT(display, K_CONTROL, is_press);
-	if (flags & MOD_SHIFT) X_KEY_EVENT_WAIT(display, K_SHIFT, is_press);
-
-	X_KEY_EVENT(display, code, is_press);
-#endif
 }
 
 void tapKeyCode(MMKeyCode code, MMKeyFlags flags)
@@ -170,22 +64,10 @@ void toggleKey(char c, const bool down, MMKeyFlags flags)
 {
 	MMKeyCode keyCode = keyCodeForChar(c);
 
-	//Prevent unused variable warning for Mac and Linux.
-#if defined(IS_WINDOWS)
-	int modifiers;
-#endif
-
 	if (isupper(c) && !(flags & MOD_SHIFT)) {
 		flags |= MOD_SHIFT; /* Not sure if this is safe for all layouts. */
 	}
 
-#if defined(IS_WINDOWS)
-	modifiers = keyCode >> 8; // Pull out modifers.
-	if ((modifiers & 1) != 0) flags |= MOD_SHIFT; // Uptdate flags from keycode modifiers.
-    if ((modifiers & 2) != 0) flags |= MOD_CONTROL;
-    if ((modifiers & 4) != 0) flags |= MOD_ALT;
-    keyCode = keyCode & 0xff; // Mask out modifiers.
-#endif
 	toggleKeyCode(keyCode, down, flags);
 }
 
@@ -195,7 +77,6 @@ void tapKey(char c, MMKeyFlags flags)
 	toggleKey(c, false, flags);
 }
 
-#if defined(IS_MACOSX)
 void toggleUnicodeKey(unsigned long ch, const bool down)
 {
 	/* This function relies on the convenient
@@ -229,9 +110,6 @@ void toggleUniKey(char c, const bool down)
 {
 	toggleUnicodeKey(c, down);
 }
-#else
-	#define toggleUniKey(c, down) toggleKey(c, down, MOD_NONE)
-#endif
 
 static void tapUniKey(char c)
 {
@@ -272,14 +150,8 @@ void typeString(const char *str)
 			n = ((c & 0x07) << 18) | (c1 << 12) | (c2 << 6) | c3;
 		}
 
-		#if defined(IS_MACOSX)
 		toggleUnicodeKey(n, true);
 		toggleUnicodeKey(n, false);
-		#else
-		toggleUniKey(n, true);
-		toggleUniKey(n, false);
-		#endif
-
 	}
 }
